requirements.txt
numpy
pinecone-client
matplotlib
sentence-transformers
